requirements_clean.txt
accelerate==1.9.0
acres==0.5.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
etelemetry==0.3.1
faiss-cpu==1.11.0.post1
filelock==3.18.0
fitz==0.0.1.dev2
fsspec==2025.7.0
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pdfminer.six==20250506
peft==0.16.0
pillow==11.3.0
prov==2.1.1
psutil==7.0.0
puremagic==1.30
pycparser==2.22
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
